CP-54351 move lvmcache from /opt
